feat: holy code optimisation
