+++ src/lib/components/AssessmentDetails.svelte
@@ -98,4 +98,21 @@
       </div>
     {/if}
   </div>
-</div> 
+</div>
+
+<style>
+  @keyframes fade-in {
+    from {
+      opacity: 0;
+      transform: translateY(10px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+
+  .animate-fade-in {
+    animation: fade-in 0.3s ease-out;
+  }
+</style>
