+++ src/lib/components/AssessmentOverview.svelte
@@ -48,9 +48,26 @@
             This assessment doesn't have a detailed description yet.
           </p>
         </div>
-      {/if}
-    </div>
+          {/if}
   </div>
+</div>
+
+<style>
+  @keyframes fade-in {
+    from {
+      opacity: 0;
+      transform: translateY(10px);
+    }
+    to {
+      opacity: 1;
+      transform: translateY(0);
+    }
+  }
+
+  .animate-fade-in {
+    animation: fade-in 0.3s ease-out;
+  }
+</style>
   
   {#if assessmentData.resources?.length}
     <div
